+++ src/components/Alignment.tsx
@@ -3,14 +3,15 @@ import StyledAlignment from '@styles/components/Alignment.styled';
 
 type AlignmentProps = {
   readonly children: React.ReactNode;
+  readonly overlay?: boolean;
   readonly vertical?: boolean;
   readonly horizontal?: boolean;
 };
 
 const Alignment: React.FC<AlignmentProps> =
-  ({ children, vertical, horizontal }: AlignmentProps): JSX.Element =>
-    <StyledAlignment vertical={ vertical } horizontal={ horizontal }>
-      <div>{ children }</div>
+  ({ children, overlay, vertical, horizontal }: AlignmentProps): JSX.Element =>
+    <StyledAlignment overlay={overlay} vertical={vertical} horizontal={horizontal}>
+      <div>{children}</div>
     </StyledAlignment>;
 
 export default Alignment;