+++ docs/changelog.md
@@ -10,6 +10,7 @@ Sprint 13 / May 24, 2020
 - removed `isLogout` logic in from Login page.
 - removed `isAuthorized` from Login page.
 - removed Prompt in favor of custom Alert component.
+- refactored Alignment component to accept `overlay` prop.
 
 <hr />